fix bugs associated with merging
