+++ helping_hands_rl_envs/planners/close_loop_block_in_bowl_planner.py
@@ -1,8 +1,6 @@
 import numpy as np
-import scipy
-from helping_hands_rl_envs.simulators import constants
 from helping_hands_rl_envs.planners.close_loop_planner import CloseLoopPlanner
-from helping_hands_rl_envs.simulators.pybullet.utils import transformations
+from helping_hands_rl_envs.pybullet.utils import transformations
 
 class CloseLoopBlockInBowlPlanner(CloseLoopPlanner):
   def __init__(self, env, config):
