+++ helping_hands_rl_envs/planners/close_loop_block_pushing_planner.py
@@ -1,8 +1,6 @@
 import numpy as np
 import scipy
-from helping_hands_rl_envs.simulators import constants
 from helping_hands_rl_envs.planners.close_loop_planner import CloseLoopPlanner
-from helping_hands_rl_envs.simulators.pybullet.utils import transformations
 
 class CloseLoopBlockPushingPlanner(CloseLoopPlanner):
   def __init__(self, env, config):
