+++ helping_hands_rl_envs/planners/close_loop_household_picking_cluttered_planner.py
@@ -1,8 +1,8 @@
 import numpy as np
 
-from helping_hands_rl_envs.simulators import constants
+from helping_hands_rl_envs.pybullet.utils import constants
 from helping_hands_rl_envs.planners.close_loop_planner import CloseLoopPlanner
-from helping_hands_rl_envs.simulators.pybullet.utils import transformations
+from helping_hands_rl_envs.pybullet.utils import transformations
 
 class CloseLoopHouseholdPickingClutteredPlanner(CloseLoopPlanner):
   def __init__(self, env, config):
