+++ helping_hands_rl_envs/pybullet/equipments/tray.py
@@ -10,7 +10,7 @@ class Tray:
     self.root_dir = os.path.dirname(helping_hands_rl_envs.__file__)
     self.id = None
 
-  def initialize(self, pos=(0,0,0), rot=(0,0,0,1), size=(0.2, 0.2, 0.2), color=[0.8, 0.8, 0.8, 1]):
+  def initialize(self, pos=(0, 0, 0), rot=(0, 0, 0, 1), size=(0.2, 0.2, 0.2), color=[0.8, 0.8, 0.8, 1], transparent=False):
     '''
     :param pos:
     :param rot:
@@ -20,58 +20,151 @@ class Tray:
              Box: inclination 90
              Plate: inclination 0
     '''
-    inclination = np.pi * (45/180)
-    cos_offset = np.cos(inclination)
-    sin_offset = np.sin(inclination)
-    half_wall_height = size[2] / (2 * sin_offset)
-    half_thickness = 0.005
-    size0 = size[0] / 2 + 2 * half_wall_height * cos_offset
-    size1 = size[1] / 2 + 2 * half_wall_height * cos_offset
-    bottom_visual = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[size0, size1, half_thickness], rgbaColor=color)
-    bottom_collision = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0, size1, half_thickness])
+    tray_shrink = 0.03
+    slopes_d = 0.03
+    botton_half_thick = 0.05
+    half_thickness = 0.001
+    pos[2] -= botton_half_thick - 2 * half_thickness
+    color_inner = [0.9, 0.6, 0.6, 1]
+    if transparent:
+      color[3] = 0
+      color_inner[3] = 0
+    size[0] -= tray_shrink
+    size[1] -= tray_shrink
+    size_inner = size.copy()
+    size_outer = size.copy()
+    size_outer[0] += slopes_d
+    size_outer[1] += slopes_d
+    inclination_inner = np.pi * (70 / 180)
+    inclination_outer = np.pi * (45 / 180)
+    cos_offset_inner = np.cos(inclination_inner)
+    sin_offset_inner = np.sin(inclination_inner)
+    cos_offset_outer = np.cos(inclination_outer)
+    sin_offset_outer = np.sin(inclination_outer)
+    h_inner = (size[2] / np.tan(inclination_outer) + slopes_d / 2) / np.tan(inclination_inner)
+    size_inner[2] = h_inner
+    half_wall_height_inner = size_inner[2] / (2 * cos_offset_inner)
+    half_wall_height_outer = size_outer[2] / (2 * cos_offset_outer)
+    size0_inner = size_inner[0] / 2 + 2 * half_wall_height_inner * sin_offset_inner
+    size1_inner = size_inner[1] / 2 + 2 * half_wall_height_inner * sin_offset_inner
+    size0_outer = size_outer[0] / 2 + 2 * half_wall_height_outer * sin_offset_outer
+    size1_outer = size_outer[1] / 2 + 2 * half_wall_height_outer * sin_offset_outer
+    bottom_visual = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[size0_inner, size1_inner, botton_half_thick],
+                                         rgbaColor=color)
+    bottom_collision = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0_inner, size1_inner, botton_half_thick])
 
-    front_visual = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1, half_wall_height], rgbaColor=color)
-    front_collision = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1, half_wall_height])
+    front_visual_inner = pb.createVisualShape(pb.GEOM_BOX,
+                                              halfExtents=[half_thickness, size1_inner, half_wall_height_inner],
+                                              rgbaColor=color_inner)
+    front_collision_inner = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1_inner,
+                                                                              half_wall_height_inner])
 
-    back_visual = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1, half_wall_height], rgbaColor=color)
-    back_collision = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1, half_wall_height])
+    back_visual_inner = pb.createVisualShape(pb.GEOM_BOX,
+                                             halfExtents=[half_thickness, size1_inner, half_wall_height_inner],
+                                             rgbaColor=color_inner)
+    back_collision_inner = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1_inner,
+                                                                             half_wall_height_inner])
 
-    left_visual = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[size0, half_thickness, half_wall_height], rgbaColor=color)
-    left_collision = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0, half_thickness, half_wall_height])
+    left_visual_inner = pb.createVisualShape(pb.GEOM_BOX,
+                                             halfExtents=[size0_inner, half_thickness, half_wall_height_inner],
+                                             rgbaColor=color_inner)
+    left_collision_inner = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0_inner, half_thickness,
+                                                                             half_wall_height_inner])
 
-    right_visual = pb.createVisualShape(pb.GEOM_BOX, halfExtents=[size0, half_thickness, half_wall_height], rgbaColor=color)
-    right_collision = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0, half_thickness, half_wall_height])
+    right_visual_inner = pb.createVisualShape(pb.GEOM_BOX,
+                                              halfExtents=[size0_inner, half_thickness, half_wall_height_inner],
+                                              rgbaColor=color_inner)
+    right_collision_inner = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0_inner, half_thickness,
+                                                                              half_wall_height_inner])
 
+    front_visual_outer = pb.createVisualShape(pb.GEOM_BOX,
+                                              halfExtents=[half_thickness, size1_outer, half_wall_height_outer],
+                                              rgbaColor=color)
+    front_collision_outer = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1_outer,
+                                                                              half_wall_height_outer])
+
+    back_visual_outer = pb.createVisualShape(pb.GEOM_BOX,
+                                             halfExtents=[half_thickness, size1_outer, half_wall_height_outer],
+                                             rgbaColor=color)
+    back_collision_outer = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[half_thickness, size1_outer,
+                                                                             half_wall_height_outer])
+
+    left_visual_outer = pb.createVisualShape(pb.GEOM_BOX,
+                                             halfExtents=[size0_outer, half_thickness, half_wall_height_outer],
+                                             rgbaColor=color)
+    left_collision_outer = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0_outer, half_thickness,
+                                                                             half_wall_height_outer])
+
+    right_visual_outer = pb.createVisualShape(pb.GEOM_BOX,
+                                              halfExtents=[size0_outer, half_thickness, half_wall_height_outer],
+                                              rgbaColor=color)
+    right_collision_outer = pb.createCollisionShape(pb.GEOM_BOX, halfExtents=[size0_outer, half_thickness,
+                                                                              half_wall_height_outer])
 
     self.id = pb.createMultiBody(baseMass=0,
                                  baseCollisionShapeIndex=bottom_collision,
                                  baseVisualShapeIndex=bottom_visual,
                                  basePosition=pos,
                                  baseOrientation=rot,
-                                 linkMasses=[1, 1, 1, 1],
-                                 linkCollisionShapeIndices=[front_collision, back_collision, left_collision, right_collision],
-                                 linkVisualShapeIndices=[front_visual, back_visual, left_visual, right_visual],
-                                 linkPositions=[[-size[0]/2 - cos_offset * size[2]/2, 0, size[2]/2],
-                                                [ size[0]/2 + cos_offset * size[2]/2, 0, size[2]/2],
-                                                [0, -size[1]/2 - cos_offset * size[2]/2, size[2]/2],
-                                                [0,  size[1]/2 + cos_offset * size[2]/2, size[2]/2]],
-                                 linkOrientations=[pb.getQuaternionFromEuler([0., -inclination, 0.]),
-                                                   pb.getQuaternionFromEuler([ 0., inclination, 0.]),
-                                                   pb.getQuaternionFromEuler([ inclination, 0., 0.]),
-                                                   pb.getQuaternionFromEuler([-inclination, 0., 0.])],
-                                 linkInertialFramePositions=[[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0]],
-                                 linkInertialFrameOrientations=[[0, 0, 0, 1], [0, 0, 0, 1], [0, 0, 0, 1], [0, 0, 0, 1]],
-                                 linkParentIndices=[0, 0, 0, 0],
-                                 linkJointTypes=[pb.JOINT_FIXED, pb.JOINT_FIXED, pb.JOINT_FIXED, pb.JOINT_FIXED],
-                                 linkJointAxis=[[0, 0, 1], [0, 0, 1], [0, 0, 1], [0, 0, 1]]
-    )
+                                 linkMasses=[1, 1, 1, 1, 1, 1, 1, 1],
+                                 linkCollisionShapeIndices=[front_collision_inner, back_collision_inner,
+                                                            left_collision_inner, right_collision_inner,
+                                                            front_collision_outer, back_collision_outer,
+                                                            left_collision_outer, right_collision_outer],
+                                 linkVisualShapeIndices=[front_visual_inner, back_visual_inner,
+                                                         left_visual_inner, right_visual_inner,
+                                                         front_visual_outer, back_visual_outer,
+                                                         left_visual_outer, right_visual_outer],
+                                 linkPositions=[[-size_inner[0] / 2 - sin_offset_inner * half_wall_height_inner, 0,
+                                                 size_inner[2] / 2 + botton_half_thick],
+                                                [size_inner[0] / 2 + sin_offset_inner * half_wall_height_inner, 0,
+                                                 size_inner[2] / 2 + botton_half_thick],
+                                                [0, -size_inner[1] / 2 - sin_offset_inner * half_wall_height_inner,
+                                                 size_inner[2] / 2 + botton_half_thick],
+                                                [0, size_inner[1] / 2 + sin_offset_inner * half_wall_height_inner,
+                                                 size_inner[2] / 2 + botton_half_thick],
+                                                [-size_outer[0] / 2 - sin_offset_outer * half_wall_height_outer, 0,
+                                                 size_outer[2] / 2 + botton_half_thick],
+                                                [size_outer[0] / 2 + sin_offset_outer * half_wall_height_outer, 0,
+                                                 size_outer[2] / 2 + botton_half_thick],
+                                                [0, -size_outer[1] / 2 - sin_offset_outer * half_wall_height_outer,
+                                                 size_outer[2] / 2 + botton_half_thick],
+                                                [0, size_outer[1] / 2 + sin_offset_outer * half_wall_height_outer,
+                                                 size_outer[2] / 2 + botton_half_thick]],
+                                 linkOrientations=[pb.getQuaternionFromEuler([0., -inclination_inner, 0.]),
+                                                   pb.getQuaternionFromEuler([0., inclination_inner, 0.]),
+                                                   pb.getQuaternionFromEuler([inclination_inner, 0., 0.]),
+                                                   pb.getQuaternionFromEuler([-inclination_inner, 0., 0.]),
+                                                   pb.getQuaternionFromEuler([0., -inclination_outer, 0.]),
+                                                   pb.getQuaternionFromEuler([0., inclination_outer, 0.]),
+                                                   pb.getQuaternionFromEuler([inclination_outer, 0., 0.]),
+                                                   pb.getQuaternionFromEuler([-inclination_outer, 0., 0.])],
+                                 linkInertialFramePositions=[[0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0],
+                                                             [0, 0, 0], [0, 0, 0], [0, 0, 0], [0, 0, 0]],
+                                 linkInertialFrameOrientations=[[0, 0, 0, 1], [0, 0, 0, 1], [0, 0, 0, 1],
+                                                                [0, 0, 0, 1],
+                                                                [0, 0, 0, 1], [0, 0, 0, 1], [0, 0, 0, 1],
+                                                                [0, 0, 0, 1]],
+                                 linkParentIndices=[0, 0, 0, 0, 0, 0, 0, 0],
+                                 linkJointTypes=[pb.JOINT_FIXED, pb.JOINT_FIXED, pb.JOINT_FIXED, pb.JOINT_FIXED,
+                                                 pb.JOINT_FIXED, pb.JOINT_FIXED, pb.JOINT_FIXED, pb.JOINT_FIXED],
+                                 linkJointAxis=[[0, 0, 1], [0, 0, 1], [0, 0, 1], [0, 0, 1],
+                                                [0, 0, 1], [0, 0, 1], [0, 0, 1], [0, 0, 1]]
+                                 )
+    for i in range(1, 9):
+      pb.changeDynamics(self.id,
+                        i,
+                        lateralFriction=0.,
+                        rollingFriction=0.,
+                        linearDamping=0.1)
 
     pb.changeDynamics(self.id,
                       -1,
+                      lateralFriction=0.5,
                       rollingFriction=0.01,
                       linearDamping=0.1)
 
-  def reset(self, pos=(0,0,0), rot=(0,0,0,1)):
+  def reset(self, pos=(0, 0, 0), rot=(0, 0, 0, 1)):
     pb.resetBasePositionAndOrientation(self.id, pos, rot)
 
   def remove(self):
