+++ helping_hands_rl_envs/pybullet/objects/bowl.py
@@ -40,6 +40,8 @@ class Bowl(PybulletObject):
 
     super(Bowl, self).__init__(constants.BOWL, object_id)
 
+    pb.changeVisualShape(object_id, -1, rgbaColor=[1, 0, 0, 1])
+
   def getGraspRotation(self):
     return self.getGraspPose()[1]
