+++ tests/test_bullet_close_loop_block_pushing.py
@@ -12,18 +12,18 @@ class TestBulletBowlStacking(unittest.TestCase):
                           [-0.15, 0.15],
                           [0.01, 0.25]])
   env_config = {'workspace': workspace, 'max_steps': 20, 'obs_size': 128, 'render': False, 'fast_mode': True,
-                'seed': 0, 'action_sequence': 'pxyz', 'num_objects': 1, 'random_orientation': False,
+                'seed': 0, 'action_sequence': 'pxyzr', 'num_objects': 1, 'random_orientation': True,
                 'reward_type': 'sparse', 'simulate_grasp': True, 'perfect_grasp': False, 'robot': 'panda',
                 'workspace_check': 'point', 'physics_mode': 'fast', 'hard_reset_freq': 1000, 'object_scale_range': (1, 1),
                 'view_type': 'render_center_height', 'transparent_bin': False, 'collision_penalty': False}
 
-  planner_config = {'random_orientation': False, 'dpos': 0.05, 'drot': np.pi/4}
+  planner_config = {'random_orientation': True, 'dpos': 0.05, 'drot': np.pi/4}
 
   def testPlanner2(self):
     self.env_config['render'] = True
     self.env_config['seed'] = 0
     num_processes = 1
-    env = env_factory.createEnvs(num_processes, 'pybullet', 'close_loop_block_pushing', self.env_config, self.planner_config)
+    env = env_factory.createEnvs(num_processes, 'close_loop_block_pushing', self.env_config, self.planner_config)
     total = 0
     s = 0
     step_times = []
@@ -33,8 +33,8 @@ class TestBulletBowlStacking(unittest.TestCase):
       t0 = time.time()
       action = env.getNextAction()
       t_plan = time.time() - t0
-      plt.imshow(obs[0, 0])
-      plt.show()
+      # plt.imshow(obs[0, 0])
+      # plt.show()
 
       # (states_, in_hands_, obs_), rewards, dones = env.simulate(action)
       # plt.imshow(obs_[0, 0])
